Move `QueryHistoryInfo` to separate file
